Reorganize global config
